restructured router
