+++ server/controllers/lobby.js
@@ -1,0 +1,10 @@
+/* signup routes */
+var express      = require('express'),
+  router         = express.Router(),
+  printlog       = require('./../helpers').printlog;
+
+router.get('/', function (req, res) {
+  res.render('lobby', {});
+});
+
+module.exports = router;
