+++ server/server.js
@@ -4,9 +4,12 @@ var express  = require('express'),
   path       = require('path'),
   db         = require('./database').db, // ./database is the relative path
   session    = require('express-session'),
+  printlog   = require('./helpers').printlog,
   // server variables
   app        = express(),
+  appRouter  = require('./controllers'),
   env        = process.env.NODE_ENV || 'production',
+  port       = process.env.PORT || 3000,
   sess;
 
 app.set('env', env);
@@ -35,106 +38,12 @@ if (app.get('env') === 'production') {
   sess.cookie.secure = true // serve secure cookies
 }
 
+app.use(appRouter);
 app.use(session(sess))
 
 // already created
 // db.createTable({name: 'users'});
 
-app.get('/', function (req, res) {
-  res.render('home', {
-    user: {
-      isAdmin: true,
-      username: 'admin'
-    }
-  });
-});
-
-app.get('/lobby', function (req, res) {
-  res.render('lobby', {});
-});
-
-function printlog (str) {
-  var isDevMode = process.env.NODE_ENV === 'development';
-  if (true || isDevMode) {
-    // do not change this to printlog() !!!!!
-    console.log(str);
-  }
-}
-
-app.get('/admin', function (req, res) {
-  var rows = [],
-    i,
-    offset = new Date().getTimezoneOffset() / -60;
-
-    // TODO - move this under --mockdata flag when ready
-    res.render('admin', {rows: [
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-      {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false}
-    ]});
-
-  // db.getTable('users')
-  //   .then(function (rows) {
-  //     printlog(rows);
-  //     // I don't like looping through records here :(
-  //     for (i = 0, len = rows.length; i < len; i++) {
-  //       var unixval = parseInt(rows[i].lastlogin, 10);
-
-  //       // Stackoverflow - http://stackoverflow.com/questions/11124322/get-date-time-for-a-specific-time-zone-using-javascript
-  //       var today = new Date(unixval + offset * 3600 * 1000).toUTCString().replace( / GMT$/, "" );
-
-  //       rows[i].lastlogin = today.toString();
-  //     }
-  //     res.render('admin', {rows: rows});
-  //   })
-  //   .catch(function (err) {
-  //     res.render('admin', {rows: [
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false},
-  //       {email: 'sam@test.com', password: 'password', lastlogin: 'Jane 4, 2017, Monday', isadmin: false}
-  //     ]});
-  //     printlog('getTable() error ---> ' + err);
-  //   });
-});
-
-app.post('/login', function (req, res) {
-  var body = req.body || {};
-  if (!Object.keys(body).length) {
-    return;
-  }
-
-  db.addUser({
-    email: body.email || '',
-    password: body.pwd || '',
-    isAdmin: false,
-    table: 'users'
-  });
-
-  // redirect to lobby after user has logged in
-  res.render('lobby', {});
-});
-
-app.get('/signup', function (req, res) {
-    res.render('signup', {
-
-    });
-});
-
-var port = process.env.PORT || 3000;
 app.listen(port, function() {
   printlog('Server started on port ' + port);
   if (process.env.NODE_ENV === 'development') {
